third_party/blink/web_tests/external/wpt/html/semantics/interactive-elements/the-details-element/toggleEvent-expected.txt

This is a testharness.js-based test.
FAIL Adding open to 'details' should fire a toggle event at the 'details' element, with 'oldState: closed' and 'newState: open' assert_equals: expected (string) "closed" but got (undefined) undefined
FAIL Adding open to 'details' and then removing open from that 'details' should fire only one toggle event at the 'details' element, with 'oldState: closed' and 'newState: closed' assert_equals: expected (string) "closed" but got (undefined) undefined
FAIL Adding open to 'details' (display:none) should fire a toggle event at the 'details' element, with 'oldState: closed' and 'newState: open' assert_equals: expected (string) "closed" but got (undefined) undefined
FAIL Adding open to 'details' (no children) should fire a toggle event at the 'details' element, with 'oldState: closed' and 'newState: open' assert_equals: expected (string) "closed" but got (undefined) undefined
FAIL Adding open to 'details' and then removing open from that 'details' and then again adding open to that 'details' should fire only one toggle event at the 'details' element, with 'oldState: closed' and 'newState: closed' assert_equals: expected (string) "closed" but got (undefined) undefined
FAIL Adding open to 'details' using setAttribute('open', '') should fire a toggle event at the 'details' element, with 'oldState: closed' and 'newState: open' assert_equals: expected (string) "closed" but got (undefined) undefined
FAIL Adding open to 'details' and then calling removeAttribute('open') should fire only one toggle event at the 'details' element, with 'oldState: closed' and 'newState: closed' assert_equals: expected (string) "closed" but got (undefined) undefined
PASS Setting open=true on an opened 'details' element should not fire a toggle event at the 'details' element
PASS Setting open=false on a closed 'details' element should not fire a toggle event at the 'details' element
FAIL Adding open to 'details' (not in the document) should fire a toggle event at the 'details' element, with 'oldState: closed' and 'newState: open' assert_equals: expected (string) "closed" but got (undefined) undefined
Harness: the test ran to completion.
